Mission briefing, JTAC stuff (library updates)
* Mission briefing now has accurate JTAC info, along with KC-130
TACAN and is now at the top of the text block for easy access.

* JTAC2 is being a bit of a jerk, and should stop lasing infrantry,
because that's a war crime.  I verified that JTAC2 is configured
correctly, and have updated the MIST and CTLD libraries which make
JTAC2 do stuff in the hope that this makes things better.  The libraries
originally included with the mission were super old.  This could break
stuff.

* Moved SA-8 SAM trigger zones a bit closer because we need more danger.